#273 created Alkis_landparcelFeatureRenderer
Created Alkis_landparcelFeatureRenderer, whose only task is to read a
hex from a config file and set it as the filling style color. If a
problem occurs with reading the value the color will be set to 'aqua
sky'. The value will be read from
/de/cismet/cids/custom/wunda_blau/res/alkis/alkis_conf.properties and
its key is LANDPARCEL_FEATURE_RENDERER_COLOR