fixed #635 adopted getters in SB_statbildserieGridObject
the stadtbildserie grid objct has a getter for the bildnummer which is
used to show the first picture in the aggregation renderer and a getter
for the bilnummer under the marker which is used to initialize the table
selection in the information panel. both getters are extended to return
the bildnummer of the vorschaubild for the first index.