issue #472
several changes (this is bad, i know...)

There was some issues with the KK editors that was added as
PropertyChangeListeners to the CidsBean, but never removed.

There was also an issue with the GeometryCombobox which dispose method
was not called in the case of the KompensationEditor => this results in a MemoryLeak

As new features, there is now a check for avoiding that null objects are added
to ausgangsbiotop/zielbiotope/massnahmen of a kompensation object.

Wordwrapping has bean enabled to the Kommentar and Nebenbestimmung Textareas.

An issue of duplicate entries for new ausgangsbiotop/zielbiotope/massnahmen objects
has been fixed.